Big fix

+++ src/main/java/edu/vsu/siuo/domains/dto/SolutionDto.java
@@ -19,7 +19,7 @@ public class SolutionDto {
     /**
      * Коэффициент удаления, dK / dCt
      */
-    private int kY;
+    private double kY;
 
     /**
      * Шаг угломера
